Add an 'anonymous activity timeout' to identity sessions.
This allows sessions that aren't for logged in people to be a different length
than those that are for logged in people.

The aim is to keep the number of identity sessions from growing too much if a
site has a lot of anonymous usage, but also a high activity timeout.

You probably want to set this to as long as might be necessary for any use of
the identity session that an anonymous user could plausably want. This could
potentially be nothing, in which case the default of 15 minutes should be fine.