Add config option: specify a regex, that if mail 'to' address matches, the mail is not sent.
I am using this because some email addresses on InstantPredict are set to a
known invalid value deliberately (because people are signing up with FB login,
but FB is not providing a valid email address for them). Since I know the
addresses are invalid, I want to prevent sending to them, but don't really want
to change any of my calls to mail.send.