Make trex-post-simpleconfirm widget work in firefox, IE etc.
Attaching dom nodes to the <button> itself only seems to work in chrome. In IE,
it just doesn't work at all. In firefox, FF assumes that any area taken up by
nodes attached to the button are actually part of the button, so you can't
click on them without it counting as a button click.