New mechanism for implementing <title> tags.
You should specify your <title> tag like this:

    <title>{{ page_title|page_title("default here") }}</title>

Then, in each template:

    {% set page_title = 'Edit widget' %}

or perhaps

    {% set page_title = 'Details for %s'|format(widget.name) %}

The default behaviour is to output a sensible page title, and CRASH if the page
title is not set - ensuring that you always will set one.

If crashing is too mainstream, you can also pass strict=false as the second
argument to the page_title filter - although it's probably better to set the
page title everywhere.