BREAKING CHANGE: Make check_login raise exceptions rather than return false
This is so that you can override it and have it raise exceptions with custom messages for whatever application-level reasons you desire, rather than it just returning false and you guessing as to the reason.

If you've overridden it previously to just return False, you probably want to change that.