No longer read the 'prop' of href for modal confirms.
The problem with the 'prop' is that when the href is '', the prop returns the
full url of the page we're on. This is probably not what is intended - normally
what you're after is just that browsers CSS-style the link as if it had a href
attr.

If you really want to programatically change the target, that's what the
data-href is for.