Update quantum self-tests to take into account changes in dateutil.
See
https://github.com/dateutil/dateutil/issues/233#issuecomment-215895151

Basically: dateutil used to parse 2013-03-01 as march 1, but due to a
change in 2.5.2, they decided to start parsing it as january 3. It's to
do with how they are trying to handle dates that look like yymmdd or
ddmmyy, etc, and they seem to think there is an obscure but valid case
where you'd want yyyyddmm. It's stupid but then relaxed parsing is a
minefield anyway.