Make the CDN serve stale versions of content to bots.
The problem: Googlebot scrapes a page, then later comes back to render
it. Between those times, we might have deployed a change, which can
cause all the CDN resources to 404.

We obviously would not want to serve a stale CDN resource to most
people, as we are interested in them getting the most current version of
content. But in the case of bots, there's no harm in them receiving
something stale. It's most likely to be alright anyway, and at least
it's not a 404.

Part of the reason for this is to see if reducing the 404 error rate for
Googlebot makes it do more comprehensive scraping of a site.