Un-pin pymongo and start using the MongoClient object in pymongo 3.x
Pymongo was pinned years ago because of a bug we found in mongoengine.
The bug was fixed and we unpinned mongoengine, but never unpinned
pymongo. Using newer pymongo will keep us compatible with newer mongo
server versions etc.

There should be no change required to your client code, since in theory
you were always using mongoengine which should have hidden pymongo
details from you. Still, the changelog is here if you want to check (we
used to be pinned to 2.6.2):
https://api.mongodb.com/python/current/changelog.html