Make the CDN serve stale files to clients, without caching.
It used to do this just for bots, but some classes of error I've been
seeing can be explained by clients attempting to access old files and
getting 404s.

Sadly the logs don't bare this theory out, but it's a change I'm willing
to try and see. Ultimately if a client requests a stale URL, it probably
won't be requesting it for that long before it works out there's a new
URL anyway. And if it really is requesting something super old, it can
have the new content and be damned, it's better than nothing.