Add a "passthrough" form of URLs in CSS files compiled by the CDN.
Allow URLs in CSS files compiled by the CDN to have a "passthrough" mode
in which they're not translated to CDN URLs. This means you can go
url('passthrough:/assets/....') in your CDN-compiled CSS file, and have
the resulting url just be the raw one, allowing you to reference URLs
that aren't actually in the CDN. This can be useful on some occasions.