Fix generation of CSRF tokens for python3.
In python2, bytes() was an alias for str(), thus bytes(a ^ b) was the same as
str(int), which did what you thought it would do.

In python3, bytes(int) does something else entirely - creates a bytes object
that is the length of the int, prepopulated with nulls.

However, bytes(seq) will create a bytes object initialised with the ints in the
sequence. Thus, wrap the `a ^ b` part as a tuple:

    bytes((a ^ b,))

There might be a more elegant solution, but I'm not looking for it right now.